+++ src/sin_code_bundle/__init__.py
@@ -3,4 +3,4 @@
 Docs: __init__.doc.md
 """
 
-__version__ = "0.6.4"
+__version__ = "0.6.5"